Output QType needs to be set as `torch.float16` for KV Cache (#2234)
The qscale type is being used to infer the q_output dtype which results
in `torch.float32` while `torch.float16` is correct. Hardcoded for now
but eventually `attn_dtype` should be plumbed.

This occurs when exporting with the torch operator and using an `fp8`
cache but dequant to `fp16` for attention. Guarantee correct attention
dtypes avoids running the `torch` operator with mixed types.

+++ sharktank/sharktank/layers/paged_attention.py
@@ -232,8 +232,8 @@ class DefaultPagedKVCache(PagedKVCache):
         key = key.transpose(2, 3).flatten(1, 2)
         value = value.transpose(2, 3).flatten(1, 2)
 
-        key = pack_raw_tensor(key, k_quantizer)
-        value = pack_raw_tensor(value, v_quantizer)
+        key = pack_raw_tensor(key, k_quantizer, dtype=torch.float16)
+        value = pack_raw_tensor(value, v_quantizer, dtype=torch.float16)
 
         return key, value
