+++ sharktank/sharktank/ops/default_impls.py
@@ -455,7 +455,11 @@ def expand_quantized(tensor: QuantizedTensor, shape: List[int]) -> QuantizedTens
     if isinstance(unpacked, TensorScaledLayout):
         new_qs = unpacked._qs.expand(*shape)
         layout = TensorScaledLayout(
-            shape=new_qs.shape, d=unpacked._d, qs=new_qs, m=unpacked._m
+            shape=new_qs.shape,
+            d=unpacked._d,
+            qs=new_qs,
+            m=unpacked._m,
+            dtype=unpacked.dtype,
         )
         return PlanarQuantizedTensor(shape=new_qs.shape, layout=layout)
     return NotImplemented
@@ -476,7 +480,11 @@ def flatten_quantized(
     if isinstance(unpacked, TensorScaledLayout):
         new_qs = torch.flatten(unpacked._qs, start_dim, end_dim)
         layout = TensorScaledLayout(
-            shape=new_qs.shape, d=unpacked._d, qs=new_qs, m=unpacked._m
+            shape=new_qs.shape,
+            d=unpacked._d,
+            qs=new_qs,
+            m=unpacked._m,
+            dtype=unpacked.dtype,
         )
         return PlanarQuantizedTensor(shape=new_qs.shape, layout=layout)
     return NotImplemented
@@ -1013,7 +1021,11 @@ def unsqueeze_quantized(tensor: QuantizedTensor, dim: int) -> QuantizedTensor:
     if isinstance(unpacked, TensorScaledLayout):
         new_qs = unpacked._qs.unsqueeze(dim)
         layout = TensorScaledLayout(
-            shape=new_qs.shape, d=unpacked._d, qs=new_qs, m=unpacked._m
+            shape=new_qs.shape,
+            d=unpacked._d,
+            qs=new_qs,
+            m=unpacked._m,
+            dtype=unpacked.dtype,
         )
         return PlanarQuantizedTensor(shape=new_qs.shape, layout=layout)
     return NotImplemented
